pyproject.toml
[project]
name = "url2md"
version = "0.1.0"
description = "URL analysis and classification tool - Generate Markdown reports from URLs"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "url2md contributors"}
]
license = {text = "CC0-1.0"}
keywords = ["url", "markdown", "analysis", "classification", "ai", "summarization"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: CC0 1.0 Universal (CC0 1.0) Public Domain Dedication",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP",
    "Topic :: Text Processing :: Markup :: Markdown",
    "Topic :: Scientific/Engineering :: Information Analysis",
]

dependencies = [
    "google-genai>=1.19.0",
    "requests>=2.31.0",
    "playwright>=1.52.0",
    "tqdm>=4.67.1",
    "minify-html>=0.16.4",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
]

[project.scripts]
url2md = "url2md.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["url2md"]

[tool.black]
line-length = 88
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 88

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]